Arithematic operations

+++ 3_Implementation/src/calculator_operations.c
@@ -1,21 +1,21 @@
 #include <calculator_operations.h>
 
-int add(int operand1, int operand2)
+float add(float operand1, float operand2)
 {
     return operand1 + operand2;
 }
 
-int subtract(int operand1, int operand2)
+float subtract(float operand1, float operand2)
 {
     return operand1 - operand2;
 }
 
-int multiply(int operand1, int operand2)
+float multiply(float operand1, float operand2)
 {
     return operand1 * operand2;
 }
 
-int divide(int operand1, int operand2)
+float divide(float operand1, float operand2)
 {
     if(0 == operand2)
         return 0;
